REFACTOR: added index.tsx on components/pond for better imports
