+++ components/pond/index.tsx
@@ -1,0 +1,3 @@
+export { default as AddPond } from '@/components/pond/AddPond';
+export { default as PondCard } from '@/components/pond/PondCard';
+export { default as PondList } from '@/components/pond/PondList';
